:truck: Move to fundamentals.py to resolve circular imports
