+++ tests/test_opengraph.py
@@ -2,9 +2,9 @@ from __future__ import annotations
 
 import networkx as nx
 
+from graphix.fundamentals import Plane
 from graphix.measurements import Measurement
 from graphix.opengraph import OpenGraph
-from graphix.pauli import Plane
 
 
 # Tests whether an open graph can be converted to and from a pattern and be
